refactor(ui): remplace le composant entreprise-fiscalite par un composant tsx
