--- packages/ui/src/components/_ui/pure-loader.vue
@@ -1,42 +1,0 @@
-<template>
-  <div class="top-level" :style="data.status !== 'LOADED' ? { display: 'flex', ['justify-content']: 'center' } : ''">
-    <template v-if="data.status === 'LOADED'">
-      <slot :item="data.value" />
-    </template>
-    <div v-else-if="data.status === 'ERROR'">
-      <HelpTooltip icon="error-warning" :text="data.message" />
-    </div>
-    <div v-else class="spinner"></div>
-  </div>
-</template>
-
-<script setup lang="ts" generic="T extends any">
-import { AsyncData } from '@/api/client-rest'
-import { HelpTooltip } from '@/components/_ui/help-tooltip'
-
-defineProps<{ data: AsyncData<T> }>()
-</script>
-<style scoped>
-.top-level {
-  position: relative;
-  min-width: var(--unit);
-  min-height: var(--unit);
-}
-@keyframes spinner {
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-.spinner {
-  position: absolute;
-  box-sizing: border-box;
-  width: var(--unit);
-  height: var(--unit);
-  border-radius: 50%;
-  border: 2px solid transparent;
-  border-top-color: var(--dsfr-bf500);
-  border-bottom-color: var(--dsfr-bf500);
-  animation: spinner 0.8s ease infinite;
-}
-</style>
